requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
pandas
streamlit
python-dotenv
sentence-transformers